Add stream of blocks
