+++ crates/services/block_aggregator_api/src/api/protobuf_adapter.rs
@@ -7,6 +7,9 @@ use crate::{
     result::Result,
 };
 use async_trait::async_trait;
+use futures::StreamExt;
+use std::cell::RefCell;
+use tokio::task::JoinHandle;
 use tokio_stream::wrappers::ReceiverStream;
 use tonic::Status;
 
@@ -17,6 +20,9 @@ use crate::{
     result::Error,
 };
 use block_aggregator_server::BlockAggregator;
+
+#[cfg(test)]
+mod tests;
 
 pub struct Server {
     query_sender: tokio::sync::mpsc::Sender<BlockAggregatorQuery<BlockRangeResponse>>,
@@ -32,7 +38,6 @@ impl Server {
 
 #[async_trait]
 impl BlockAggregator for Server {
-    type GetBlockRangeStream = ReceiverStream<Result<protobuf_adapter::Block, Status>>;
     async fn get_block_height(
         &self,
         request: tonic::Request<BlockHeightRequest>,
@@ -54,12 +59,59 @@ impl BlockAggregator for Server {
             ))),
         }
     }
+    type GetBlockRangeStream = ReceiverStream<Result<protobuf_adapter::Block, Status>>;
 
     async fn get_block_range(
         &self,
         request: tonic::Request<BlockRangeRequest>,
     ) -> Result<tonic::Response<Self::GetBlockRangeStream>, tonic::Status> {
-        todo!()
+        tracing::warn!("get_block_range: {:?}", request);
+        let req = request.into_inner();
+        let (response, receiver) = tokio::sync::oneshot::channel();
+        let query = BlockAggregatorQuery::GetBlockRange {
+            first: req.start.into(),
+            last: req.end.into(),
+            response,
+        };
+        self.query_sender
+            .send(query)
+            .await
+            .map_err(|e| Status::internal(format!("Failed to send query: {}", e)))?;
+        let res = receiver.await;
+        match res {
+            Ok(block_range_response) => match block_range_response {
+                BlockRangeResponse::Literal(inner) => {
+                    let (tx, rx) =
+                        tokio::sync::mpsc::channel::<Result<Block, Status>>(16);
+
+                    // TODO: is this safe if we just drop the join handle?
+                    let _ = tokio::spawn(async move {
+                        let mut s = inner;
+                        while let Some(block) = s.next().await {
+                            // Convert your internal `blocks::Block` into the protobuf `Block`.
+                            let pb = Block {
+                                data: block.bytes().to_vec(),
+                            };
+
+                            // If the receiver side was dropped, stop forwarding.
+                            if tx.send(Ok(pb)).await.is_err() {
+                                break;
+                            }
+                        }
+                    });
+
+                    Ok(tonic::Response::new(ReceiverStream::new(rx)))
+                }
+                BlockRangeResponse::Remote(_) => {
+                    tracing::error!("Remote block range not implemented");
+                    todo!()
+                }
+            },
+            Err(e) => Err(tonic::Status::internal(format!(
+                "Failed to receive block range: {}",
+                e
+            ))),
+        }
     }
 }
 
@@ -104,115 +156,3 @@ impl BlockAggregatorApi for ProtobufAPI {
 }
 
 pub struct ProtobufClient;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::blocks::Block;
-    use block_aggregator_client::BlockAggregatorClient;
-    use bytes::Bytes;
-    use fuel_core_types::fuel_types::BlockHeight;
-    use futures::{
-        StreamExt,
-        TryStreamExt,
-    };
-
-    #[tokio::test]
-    async fn await_query__get_current_height__client_receives_expected_value() {
-        let _ = tracing_subscriber::fmt()
-            .with_max_level(tracing::Level::DEBUG)
-            .init();
-        // given
-        let path = "[::1]:50051";
-        let mut api = ProtobufAPI::new(path.to_string());
-        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-
-        // call get current height endpoint with client
-        let url = "http://[::1]:50051";
-        let mut client = BlockAggregatorClient::connect(url.to_string())
-            .await
-            .expect("could not connect to server");
-        let handle = tokio::spawn(async move {
-            tracing::info!("querying with client");
-            client
-                .get_block_height(BlockHeightRequest {})
-                .await
-                .expect("could not get height")
-        });
-
-        // when
-        tracing::info!("awaiting query");
-        let query = api.await_query().await.unwrap();
-
-        // then
-        // return response through query's channel
-        if let BlockAggregatorQuery::GetCurrentHeight { response } = query {
-            response.send(BlockHeight::new(42)).unwrap();
-        } else {
-            panic!("expected GetCurrentHeight query");
-        }
-        let res = handle.await.unwrap();
-
-        // assert client received expected value
-        assert_eq!(res.into_inner().height, 42);
-    }
-
-    #[tokio::test]
-    async fn await_query__get_block_range__client_receives_expected_value() {
-        let _ = tracing_subscriber::fmt()
-            .with_max_level(tracing::Level::DEBUG)
-            .init();
-        // given
-        let path = "[::1]:50051";
-        let mut api = ProtobufAPI::new(path.to_string());
-        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-
-        // call get current height endpoint with client
-        let url = "http://[::1]:50051";
-        let mut client = BlockAggregatorClient::connect(url.to_string())
-            .await
-            .expect("could not connect to server");
-        let request = BlockRangeRequest { start: 0, end: 1 };
-        let handle = tokio::spawn(async move {
-            tracing::info!("querying with client");
-            client
-                .get_block_range(request)
-                .await
-                .expect("could not get height")
-        });
-
-        // when
-        tracing::info!("awaiting query");
-        let query = api.await_query().await.unwrap();
-
-        // then
-        let block1 = Block::new(Bytes::from(vec![0u8; 100]));
-        let block2 = Block::new(Bytes::from(vec![1u8; 100]));
-        let list = vec![block1, block2];
-        // return response through query's channel
-        if let BlockAggregatorQuery::GetBlockRange {
-            first,
-            last,
-            response,
-        } = query
-        {
-            let stream = tokio_stream::iter(list.clone()).boxed();
-            let range = BlockRangeResponse::Literal(stream);
-            response.send(range).unwrap();
-        } else {
-            panic!("expected GetBlockRange query");
-        }
-        let response = handle.await.unwrap();
-        let expected: Vec<Vec<u8>> = list.iter().map(|b| b.bytes().to_vec()).collect();
-        let actual: Vec<Vec<u8>> = response
-            .into_inner()
-            .try_collect::<Vec<_>>()
-            .await
-            .unwrap()
-            .into_iter()
-            .map(|b| b.data.to_vec())
-            .collect();
-
-        assert_eq!(expected, actual);
-    }
-}
